Добавлен обстактный класс
