+++ src/view/filter.js
@@ -1,4 +1,4 @@
-import {createElement} from './utils.js';
+import Abstract from './abstract.js';
 
 /**
  *
@@ -12,24 +12,8 @@ const createFilterTemplate = () => (
   </ul>`
 );
 
-export default class FilterMenu {
-  constructor() {
-    this._element = null;
-  }
-
+export default class FilterMenu extends Abstract {
   getTemplate() {
     return createFilterTemplate();
-  }
-
-  getElement() {
-    if(!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
