--- src/view/load-more-film/create-show-more-button.js
@@ -1,9 +1,0 @@
-/**
- *
- * @returns Создается кнопка загрузить ещё
- */
-const createShowMoreButtonTemplate = () => (
-  '<button class="films-list__show-more">Show more</button>'
-);
-
-export {createShowMoreButtonTemplate};
