+++ src/view/menu.js
@@ -1,4 +1,4 @@
-import {createElement} from './utils.js';
+import Abstract from './abstract.js';
 /**
  *
  * @returns Создает шаблон разметки меню
@@ -15,9 +15,9 @@ const createMenuTemplate = (countWatchlist, countWatched, countFavorites) => (
   </nav>`
 );
 
-class SiteMenu {
+class SiteMenu extends Abstract{
   constructor(coutWatchlist, countWatched, countFavorites) {
-    this._element = null;
+    super();
     this._countWatchlist = coutWatchlist;
     this._countWatched = countWatched;
     this._countFavorites = countFavorites;
@@ -25,17 +25,6 @@ class SiteMenu {
 
   getTemplate() {
     return createMenuTemplate(this._countWatchlist, this._countWatched, this._countFavorites);
-  }
-
-  getElement() {
-    if(!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
