+++ src/view/user-rating/create-user-rank.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import Abstract from '../abstract.js';
 import {setUserRank} from './set-user-rank.js';
 
 /**
@@ -9,32 +9,21 @@ const createProfileTemplate = (rank) => (
   rank !== null ? `<section class="header__profile profile">
     <p class="profile__rating">${rank}</p>
     <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-  </section>` : ''
+  </section>` : '<section class="header__profile profile visually-hidden"></section>'
 );
 
-class Rank {
+class Rank extends Abstract {
   /**
    *
    * @param {number} watched число просмотренных фильмов
    */
   constructor (watched) {
-    this._element = null;
+    super();
     this._rank = setUserRank(watched);
   }
 
   getTemplate() {
     return createProfileTemplate(this._rank);
-  }
-
-  getElement() {
-    if(!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
