+++ src/view/user-rating/set-user-rank.js
@@ -1,10 +1,3 @@
-/*
-0 — блок со званием не отображается;
-от 1 до 10 — novice;
-от 11 до 20 — fan;
-от 21 и выше — movie buff.
-*/
-
 /**
  * Установит ранк в зависимости от просмотренных фильмов
  */
